% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci_bounds_mc_power.R
\name{add_ci_bounds_mc_power}
\alias{add_ci_bounds_mc_power}
\title{Calculate CI bounds on Monte Carlo}
\usage{
add_ci_bounds_mc_power(power_results, nsim, conf = 0.95)
}
\value{
Power data.frame with conf intervals
}
\description{
Calculates CI bounds for Monte Carlo power
}
\keyword{internal}
