% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliasmodel.R
\name{aliasmodel}
\alias{aliasmodel}
\title{Alias terms}
\usage{
aliasmodel(formula, power)
}
\arguments{
\item{formula}{The formula to be expanded}
}
\value{
Returns aliased model terms from formula
}
\description{
Creates alias terms for Alias-optimal designs
}
\keyword{internal}
