% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anticoef_from_delta.R
\name{anticoef_from_delta}
\alias{anticoef_from_delta}
\title{Generates Anticipated Coefficients from delta}
\usage{
anticoef_from_delta(default_coef, delta, glmfamily)
}
\arguments{
\item{default_coef}{a vector of default coefficients, from gen_anticoef}

\item{delta}{the user-input delta parameter, must be a numeric vector of length 1 or 2}

\item{glmfamily}{the user-supplied glmfamily, either a string or a glmfamily object}
}
\value{
Anticipated coefficients.
}
\description{
Generates Anticipated Coefficients from delta parameter
The logic for generating anticipated coefficients from delta varies with
glm family.
}
\keyword{internal}
