% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockingstructure.R
\name{blockingstructure}
\alias{blockingstructure}
\title{Find block sizes in column}
\usage{
blockingstructure(blocklist)
}
\arguments{
\item{blocklist}{Block list}
}
\value{
Returns the blocking structure given a vector of block ids
}
\description{
Returns blocks
}
\keyword{internal}
