% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conservative_anticoef.R
\name{calc_conservative_anticoef}
\alias{calc_conservative_anticoef}
\title{Calculate Conservative Anticipated Coefficients}
\usage{
calc_conservative_anticoef(results, effectsize)
}
\arguments{
\item{results}{The power results matrix}

\item{effectsize}{The effectsize.}
}
\value{
Vector of conservative anticipated coefficients
}
\description{
Calculate Conservative Anticipated Coefficients
}
\keyword{internal}
