% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_interaction_degrees.R
\name{calc_interaction_degrees}
\alias{calc_interaction_degrees}
\title{Calculate Interaction Degrees of Freedom}
\usage{
calc_interaction_degrees(
  design,
  model,
  contrast,
  nointercept,
  split_layers,
  split_degrees,
  split_plot_structure
)
}
\arguments{
\item{design}{The design matrix}

\item{model}{The model}

\item{contrast}{The contrast}

\item{split_layers}{The layer of split plots for each main effect term}

\item{split_degrees}{The number of degrees of freedom for each main effect term}
}
\value{
Degrees of freedom vector
}
\description{
Calculate interaction degrees of freedom for split-plot designs
}
\keyword{internal}
