% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcblocksizes.R
\name{calcblocksizes}
\alias{calcblocksizes}
\title{Calculate block sizes}
\usage{
calcblocksizes(trials, blocksize)
}
\arguments{
\item{trials}{Number of trials in design}

\item{blocksize}{The desired size of each block}
}
\value{
The blocksize vector
}
\description{
Calculate block size vector
}
\keyword{internal}
