% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcnoncentralparam.R
\name{calcnoncentralparam}
\alias{calcnoncentralparam}
\title{Calculate Non-Centrality Parameter}
\usage{
calcnoncentralparam(X, L, b, vInv = NULL)
}
\arguments{
\item{X}{The model matrix}

\item{L}{The parameter matrix/vector}

\item{b}{The anticipated coefficients}
}
\value{
The non-centrality parameter for the F test
}
\description{
Calculates the non-centrality parameter for the model matrix
}
\keyword{internal}
