% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_block_nesting.R
\name{calculate_block_nesting}
\alias{calculate_block_nesting}
\title{Determine Nesting Level of Blocks}
\usage{
calculate_block_nesting(blockgroups, blockstructure)
}
\value{
Vector of numbers indicating the hierarchy
}
\description{
Calculates if a block is fully nested within another block, and
what the highest level of nesting is. 1 indicates the block isn't nested within
another block--just within the intercept.
}
\keyword{internal}
