% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_degrees_of_freedom.R
\name{calculate_degrees_of_freedom}
\alias{calculate_degrees_of_freedom}
\title{Calculate Degrees of Freedom}
\usage{
calculate_degrees_of_freedom(
  run_matrix_processed,
  nointercept,
  model,
  contrasts
)
}
\arguments{
\item{run_matrix_processed}{The design}

\item{nointercept}{Whether the intercept term is present}
}
\value{
Vector of numbers indicating split plot layers
}
\description{
Calculates the degrees of freedom adjustment for models with random effects from Penheiro and Bates, pg. 91
}
\keyword{internal}
