% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_level_vector.R
\name{calculate_level_vector}
\alias{calculate_level_vector}
\title{Calculate level vector}
\usage{
calculate_level_vector(design, model, nointercept)
}
\arguments{
\item{design}{Design to be analyzed.}

\item{model}{The model.}

\item{nointercept}{Whether an intercept is included.}
}
\value{
A vector of levels.
}
\description{
Calculate level vector
}
\examples{
#We can pass either the output of gen_design or eval_design to plot_correlations
}
\keyword{internal}
