% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_split_columns.R
\name{calculate_split_columns}
\alias{calculate_split_columns}
\title{Calculate Split Plot Columns}
\usage{
calculate_split_columns(run_matrix_processed, blockgroups)
}
\arguments{
\item{run_matrix_processed}{The design}
}
\value{
Vector of numbers indicating split plot layers
}
\description{
Calculates which columns correspond to which layers of randomization restriction
}
\keyword{internal}
