% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_v_from_blocks.R
\name{calculate_v_from_blocks}
\alias{calculate_v_from_blocks}
\title{Calculate V from Blocks}
\usage{
calculate_v_from_blocks(
  nrow_design,
  blockgroups,
  blockstructure,
  varianceratios
)
}
\arguments{
\item{nrow_design}{The number of runs in the design}

\item{blockgroups}{List indicating the size of each block for each layer}

\item{blockstructure}{Matrix indicating the block structure from the rownames}
}
\value{
Variance-Covariance Matrix
}
\description{
Calculates the V matrix from the block structure
}
\keyword{internal}
