% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_suggest_packages.R
\name{check_for_suggest_packages}
\alias{check_for_suggest_packages}
\title{check_for_suggest_packages}
\usage{
check_for_suggest_packages(packages)
}
\value{
none
}
\description{
checks for suggests
}
\keyword{internal}
