% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRunMatrix.R
\name{constructRunMatrix}
\alias{constructRunMatrix}
\title{Construct Run Matrix given rows}
\usage{
constructRunMatrix(rowIndices, candidateList, augment = NULL)
}
\value{
Returns a vector consisting of the number
of levels preceding each parameter (including the intercept)
}
\description{
Returns number of levels prior to each parameter
}
\keyword{internal}
