% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.simplex.R
\name{contr.simplex}
\alias{contr.simplex}
\title{Simplex Contrast Generator}
\usage{
contr.simplex(n, size = 1)
}
\arguments{
\item{n}{The number of levels in the catagorical variable}

\item{size}{The length of the simplex vector. Default 1.}
}
\description{
Evaluates power for a right censored survival design with a Monte Carlo simulation,
using the survival package and survreg to fit the data.
}
\examples{
contr.simplex(4)
}

