% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_blockcolumn_rownames.R
\name{convert_blockcolumn_rownames}
\alias{convert_blockcolumn_rownames}
\title{Convert Block Column to Rownames}
\usage{
convert_blockcolumn_rownames(
  RunMatrix,
  blocking,
  varianceratios,
  user_specified_varianceratio,
  verbose = FALSE
)
}
\arguments{
\item{RunMatrix}{The run matrix}

\item{blocking}{Whether random effects should be included.}

\item{varianceratios}{A vector of variance ratios for each level of restricted randomization}
}
\value{
Row-name encoded blocked run matrix
}
\description{
Detect externally generated blocking columns and convert to rownames
}
\keyword{internal}
