% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_model_dots.R
\name{convert_model_dots}
\alias{convert_model_dots}
\title{Converts dot operator to terms}
\usage{
convert_model_dots(design, model, splitplotdesign = NULL)
}
\arguments{
\item{design}{The design}

\item{model}{Base model}

\item{splitplotdesign}{split plot design data.frame}
}
\value{
New model with dot operator replaced
}
\description{
Converts the dot operator \code{.} in a formula to the linear terms in the model. Includes interactions (e.g. .*.)
}
\keyword{internal}
