% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rownames_to_covariance.R
\name{convert_rownames_to_covariance}
\alias{convert_rownames_to_covariance}
\title{Converts dot operator to terms}
\usage{
convert_rownames_to_covariance(
  run_matrix_processed,
  varianceratios,
  user_specified_varianceratio
)
}
\arguments{
\item{run_matrix_processed}{The design}
}
\value{
Variance-covariance matrix V
}
\description{
Converts the row names to a variance-covariance matrix (using the user-supplied variance ratio)
}
\keyword{internal}
