% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_convex_hull_moment_matrix.R
\name{convhull_halfspace}
\alias{convhull_halfspace}
\title{Compute convex hull in half-space form}
\usage{
convhull_halfspace(points)
}
\arguments{
\item{points}{A matrix or data frame of numeric coordinates, size n x d}
}
\value{
A list with elements:
  - A: a matrix of size m x d (m = number of facets)
  - b: a length-m vector
  so that x is inside the hull if and only if A %*% x + b >= 0 (for all rows).
  - volume: Total volume of the convex hull
}
\description{
Compute convex hull in half-space form
}
\keyword{internal}
