% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUI_code_pane.R
\name{create_code_pane_reactives}
\alias{create_code_pane_reactives}
\title{Create Code Pane Reactives}
\usage{
create_code_pane_reactives(input, update, candidatesetall, any_htc_fn)
}
\arguments{
\item{input}{Shiny input object.}

\item{update}{Reactive dependency used to trigger recomputation.}

\item{candidatesetall}{Reactive that returns the full candidate set list.}

\item{any_htc_fn}{Function returning whether any factors are hard-to-change.}
}
\value{
List of reactives required exclusively by the code pane.
}
\description{
Create Code Pane Reactives
}
\keyword{internal}
