% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disallowed_combinations.R
\name{detect_disallowed_combinations}
\alias{detect_disallowed_combinations}
\title{Detects disallowed combinations in candidate set}
\usage{
detect_disallowed_combinations(candidateset)
}
\arguments{
\item{candidateset}{Data frame}
}
\value{
Returns a logical
}
\description{
Detects disallowed combinations in candidate set
}
\keyword{internal}
