% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinefactorialsize.R
\name{determinefactorialsize}
\alias{determinefactorialsize}
\title{Determines full factorial size}
\usage{
determinefactorialsize(RunMatrix)
}
\arguments{
\item{RunMatrix}{The run matrix}
}
\value{
Returns number of runs in full factorial
}
\description{
Determines full factorial size
}
\keyword{internal}

