% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disallowed_combinations.R
\name{disallowed_combinations}
\alias{disallowed_combinations}
\title{Detect and list disallowed combinations in candidate set}
\usage{
disallowed_combinations(candidateset)
}
\arguments{
\item{candidateset}{Block list}
}
\value{
Returns a list of the disallowed combinations
}
\description{
Detects and list disallowed combinations in candidate set
}
\keyword{internal}
