% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectpowermc.R
\name{effectpowermc}
\alias{effectpowermc}
\title{Fit Anova for Effect Power Calculation in Monte Carlo}
\usage{
effectpowermc(
  fit,
  type = "III",
  test = "Pr(>Chisq)",
  model_formula = NULL,
  firth = FALSE,
  glmfamily = "gaussian",
  effect_terms = NULL,
  RunMatrixReduced = NULL,
  method = NULL,
  contrastslist = contrastslist,
  effect_anova = FALSE,
  ...
)
}
\arguments{
\item{fit}{Fit from regression}

\item{type}{Default \code{III}}

\item{test}{Default \verb{Pr(>Chisq)}.}

\item{...}{Additional arguments to pass to car::Anova}
}
\value{
p-values
}
\description{
Calculates the p-values for the effect power calculation in Monte Carlo
}
\keyword{internal}
