% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSignificance.R
\name{effectSignificance}
\alias{effectSignificance}
\title{Compute effect powers from parameter powers}
\usage{
effectSignificance(pvals, alpha, parameter_assignment)
}
\arguments{
\item{pvals}{Vector of p-values, from a fit (e.g. lm, glm, etc.)}

\item{alpha}{Type-1 error.}

\item{parameter_assignment}{Vector that assigns each parameter to an effect.}
}
\value{
Returns a vector of 1s and 0s, with 1 at position i indicating that
effect i has at least one significant parameter. If parameter_assignment came from
attr(model.matrix, 'assign'), then the i-th parameter name is given by attr(terms(model), 'term.labels').
The returned vector has a length max(parameter_assignment) + 1.
}
\description{
Given a vector of parameter p-values, return a vector of effect p-values
}

