% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPvalues.R
\name{extractPvalues}
\alias{extractPvalues}
\title{Extract p-values from a model object}
\usage{
extractPvalues(model_fit)
}
\arguments{
\item{model_fit}{The model object from which to extract.}
}
\value{
Returns a vector of p-values. If model_fit is not a supported model type, returns NULL.
}
\description{
Extract p-values from a model object. Currently works with lm, glm, lme4, glmer,
and survreg model objects. If possible, uses the p-values reported in summary(model_fit).
If those do not exist (I'm looking at you, lme4), returns the Wald p-value:
2*pnorm(-abs(estimate / se))
}
\keyword{internal}
