% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_anticoef.R
\name{gen_anticoef}
\alias{gen_anticoef}
\title{Generates Default Anticipated Coefficients}
\usage{
gen_anticoef(modelmatrix)
}
\arguments{
\item{modelmatrix}{The model matrix}
}
\value{
Anticipated coefficients
}
\description{
Generates Default Anticipated Coefficients
}
\examples{
genparammatrix(6,2,1)
genparammatrix(6,2,1)
}

