% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_binomial_anticoef.R
\name{gen_binomial_anticoef}
\alias{gen_binomial_anticoef}
\title{Generates Binomial Anticipated Coefficients}
\usage{
gen_binomial_anticoef(anticoef, lowprob, highprob)
}
\arguments{
\item{anticoef}{Anticipated coefficeints}

\item{lowprob}{Default 0.50. The base probability}

\item{highprob}{Default 0.80. The base probability}

\item{runmatrix}{The Run Matrix}

\item{model}{Base model}

\item{contrastslist}{Contrasts}
}
\value{
Anticipated coefficients.
}
\description{
Generates Binomial Anticipated Coefficients
}
\keyword{internal}
