% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_momentsmatrix.R
\name{gen_momentsmatrix}
\alias{gen_momentsmatrix}
\title{Generates the moment matrix}
\usage{
gen_momentsmatrix(modelfactors, levelvector, classvector)
}
\value{
Returns a vector consisting of the number
of levels preceeding each parameter (including the intercept)
}
\description{
Returns number of levels prior to each parameter
}
\keyword{internal}
