% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_convex_hull_moment_matrix.R
\name{gen_momentsmatrix_continuous}
\alias{gen_momentsmatrix_continuous}
\title{Approximate continuous moment matrix over a numeric bounding box}
\usage{
gen_momentsmatrix_continuous(
  formula = ~.,
  candidate_set,
  n_samples_per_dimension
)
}
\arguments{
\item{formula}{Default `~ .`. Model formula specifying the terms.}

\item{n_samples_per_dimension}{Default `100`. Number of samples to take per dimension when interpolating inside
the convex hull.}

\item{data}{Candidate set}
}
\value{
A matrix of size `p x p` (where `p` is the number of columns in the model matrix),
  approximating the continuous moment matrix.
}
\description{
Treats the min–max range of each numeric column as a bounding box
and samples points uniformly to approximate the integral of the outer product of
the model terms, weighted by whether the point
is on the edge.
}
\keyword{internal}
