% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_momentsmatrix_ideal.R
\name{gen_momentsmatrix_ideal}
\alias{gen_momentsmatrix_ideal}
\title{Generates the moment matrix, assuming no constraints}
\usage{
gen_momentsmatrix_ideal(modelfactors, classvector)
}
\value{
Returns a vector consisting of the number
of levels preceding each parameter (including the intercept)
}
\description{
Returns number of levels prior to each parameter
}
\keyword{internal}
