% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUI_dynamic_ui_elements.R
\name{generate_block_panel}
\alias{generate_block_panel}
\title{Generate Block Panel}
\usage{
generate_block_panel(any_htc)
}
\arguments{
\item{any_htc}{Factor number}
}
\value{
Shiny UI
}
\description{
Generate Block Panel
}
\keyword{internal}
