% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUI_dynamic_ui_elements.R
\name{generate_factor_input_panel}
\alias{generate_factor_input_panel}
\title{Generate Factor Input Panel}
\usage{
generate_factor_input_panel(factor_n = 1, factor_input_cache = NULL)
}
\arguments{
\item{factor_n}{Factor number}
}
\value{
Shiny UI
}
\description{
Generate Factor Input Panel
}
\keyword{internal}
