% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUI_dynamic_ui_elements.R
\name{generate_optimality_results}
\alias{generate_optimality_results}
\title{Generate Optimality Results}
\usage{
generate_optimality_results(any_htc)
}
\arguments{
\item{any_htc}{Factor number}
}
\value{
Shiny UI
}
\description{
Generate Optimality Results
}
\keyword{internal}
