% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_groups.R
\name{get_block_groups}
\alias{get_block_groups}
\title{Calculate block structure lengths}
\usage{
get_block_groups(existing_block_structure)
}
\arguments{
\item{existing_block_structure}{character matrix from rownames, split by \code{.}}
}
\value{
List of numbers indicating split plot layer sizes
}
\description{
Calculate block structure lengths
}
\keyword{internal}
