% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moment_matrix.R
\name{get_moment_matrix}
\alias{get_moment_matrix}
\title{Gets or calculates the moment matrix, given a candidate set, a design, and a model}
\usage{
get_moment_matrix(
  design = NA,
  factors = NA,
  classvector = NA,
  candidate_set_normalized = NA,
  model = ~.,
  moment_sample_density = 10,
  high_resolution_candidate_set = NA
)
}
\value{
Returns a vector consisting of the number
of levels preceding each parameter (including the intercept)
}
\description{
Returns number of levels prior to each parameter
}
\keyword{internal}
