% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimality.R
\name{get_optimality}
\alias{get_optimality}
\title{Get optimality values}
\usage{
get_optimality(output, optimality = NULL)
}
\arguments{
\item{output}{The output of either gen_design or eval_design/eval_design_mc}

\item{optimality}{Default `NULL`. Return just the specific optimality requested.}
}
\value{
A dataframe of optimality conditions. `D`, `A`, and `G` are efficiencies (value is out of 100).
`T` is the trace of the information matrix, `E` is the minimum eigenvalue of the information matrix,
`I` is the average prediction variance, and `Alias` is the trace of the alias matrix.
}
\description{
Returns a list of optimality values (or one value in particular)
}
\examples{
#We can pass either the output of gen_design or eval_design to plot_correlations
}
