% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power_curves.R
\name{get_power_curve_output}
\alias{get_power_curve_output}
\title{Get Power Curve Warnings and Errors}
\usage{
get_power_curve_output(power_curve)
}
\arguments{
\item{power_curve}{The output from `calculate_power_curves()`}
}
\value{
A list of data.frames containing warning/error information
}
\description{
Gets the warnings and errors from `calculate_power_curves()` output.
}
\examples{
#Generate sample
\dontrun{
calculate_power_curves(trials=seq(50,150,by=20),
                      candidateset = expand.grid(x=c(-1,1),y=c(-1,1)),
                      model = ~.,
                      effectsize = list(c(0.5,0.9),c(0.6,0.9)),
                      eval_function = eval_design_mc,
                      eval_args = list(nsim = 100, glmfamily = "binomial"))
}
}
