% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_convex_hull_moment_matrix.R
\name{interpolate_convex_hull}
\alias{interpolate_convex_hull}
\title{Interpolate points within a convex hull}
\usage{
interpolate_convex_hull(points, ch_halfspace, n_samples_per_dimension = 11)
}
\arguments{
\item{points}{A numeric matrix (n x d).}

\item{ch_halfspace}{Convex hull. Output of \code{convhull_halfspace}}

\item{n_samples_per_dimension}{Number of samples per dimension (pre filtering) to take}
}
\value{
A list with:
\itemize{
\item data: The interpolated points
\item on_edge: A logical vector indicating whether the points
}
}
\description{
Interpolate points within a convex hull
}
\keyword{internal}
