% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_intralayer_interaction.R
\name{is_intralayer_interaction}
\alias{is_intralayer_interaction}
\title{Layer Interaction}
\usage{
is_intralayer_interaction(design, model, split_layers)
}
\arguments{
\item{design}{The design matrix}

\item{model}{The model}

\item{split_layers}{The layer of split plots for each main effect term}
}
\value{
List of booleans for each subplot strata layer
}
\description{
Determines if a factor is a intra-layer interaction
}
\keyword{internal}
