% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_numeric_runmatrix.R
\name{normalize_numeric_runmatrix}
\alias{normalize_numeric_runmatrix}
\title{Normalize Run Matrix}
\usage{
normalize_numeric_runmatrix(RunMatrix)
}
\arguments{
\item{RunMatrix}{The run matrix}
}
\value{
Normalized run matrix
}
\description{
Normalizes the numeric columns in the runmatrix
}
\keyword{internal}
