% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{permutations}
\alias{permutations}
\title{Permutations}
\usage{
permutations(n)
}
\arguments{
\item{n}{Number of elements to permute}
}
\value{
Matrix of permuted element ids
}
\description{
Return permutations
}
\keyword{internal}
