% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_functions.R
\name{potential_permuted_factors}
\alias{potential_permuted_factors}
\title{Find potential permuted interactions}
\usage{
potential_permuted_factors(x)
}
\arguments{
\item{x}{character vector of interaction terms}
}
\value{
character vector of potential interaction terms
}
\description{
Returns permuted interactions
}
\keyword{internal}
