% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorlevels.R
\name{priorlevels}
\alias{priorlevels}
\title{Prior levels}
\usage{
priorlevels(levelvector)
}
\value{
Returns a vector consisting of the number
of levels preceding each parameter (including the intercept)
}
\description{
Returns number of levels prior to each parameter
}
\keyword{internal}
