% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad.R
\name{quad}
\alias{quad}
\title{quadratic}
\usage{
quad(formula)
}
\arguments{
\item{formula}{The formula to be expanded}
}
\value{
Returns quadratic model from formula
}
\description{
quadratic
}
\keyword{internal}
