% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceRunMatrix.R
\name{reduceRunMatrix}
\alias{reduceRunMatrix}
\title{Remove columns not in model}
\usage{
reduceRunMatrix(RunMatrix, model)
}
\arguments{
\item{model}{The formula}

\item{ModelMatrix}{The model matrix}
}
\value{
The reduced model matrix.
}
\description{
Remove columns not in model
}
\keyword{internal}
