% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_skpr_blockcols.R
\name{remove_skpr_blockcols}
\alias{remove_skpr_blockcols}
\title{Remove skpr-generated blocking columns}
\usage{
remove_skpr_blockcols(RunMatrix)
}
\arguments{
\item{RunMatrix}{The run matrix}
}
\value{
Run Matrix
}
\description{
Remove skpr-generated REML blocking columns if present
}
\keyword{internal}
