% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_progressr_handler.R
\name{set_up_progressr_handler}
\alias{set_up_progressr_handler}
\title{Set up progressr handler}
\usage{
set_up_progressr_handler(msg_string, type_string)
}
\description{
Set up progressr handler
}
\keyword{internal}
