% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUI.R
\name{skprGUI}
\alias{skprGUI}
\title{Graphical User Interface for skpr}
\usage{
skprGUI(
  browser = FALSE,
  return_app = FALSE,
  multiuser = FALSE,
  progress = TRUE
)
}
\arguments{
\item{browser}{Default \code{FALSE}. Whether to open the application in an external browser.}

\item{return_app}{Default \code{FALSE}. If \code{TRUE}, this will return the shinyApp object.}

\item{multiuser}{Default \code{FALSE}. If \code{TRUE}, this will turn off and disable multicore functionality and enable non-blocking operation.}

\item{progress}{Default \code{TRUE}. Whether to include a progress bar in the application. Note: if \code{multiuser = TRUE}, progress
bars are turned on by default.}
}
\description{
skprGUI provides a graphical user interface to skpr, within R Studio.
}
\examples{
#Type `skprGUI()` to begin

}
