% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUIbrowser.R
\name{skprGUIbrowser}
\alias{skprGUIbrowser}
\title{skprGUIbrowser}
\usage{
skprGUIbrowser()
}
\description{
skprGUI provides a graphical user interface to skpr, in an external browser.
}
\examples{
#Type skprGUIbrowser() to begin

\dontrun{skprGUIbrowser()}
}
