% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUIserver.R
\name{skprGUIserver}
\alias{skprGUIserver}
\title{Graphical User Interface for skpr}
\usage{
skprGUIserver()
}
\description{
skprGUI provides a graphical user interface to skpr, within R Studio.
}
\examples{
#Type `skprGUIserver()` to begin

}
