% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anticoef_from_delta_surv.R
\name{anticoef_from_delta_surv}
\alias{anticoef_from_delta_surv}
\title{Generates Anticipated Coefficients from delta for eval_design_suvival_mc}
\usage{
anticoef_from_delta_surv(default_coef, delta, distribution)
}
\arguments{
\item{default_coef}{a vector of default coefficients, from gen_anticoef}

\item{delta}{the user-input delta parameter, must be a numeric vector of length 1 or 2}

\item{distribution}{the user-supplied distribution, either a string or a survreg distribution object}
}
\value{
Anticipated coefficients.
}
\description{
Generates Anticipated Coefficients from delta parameter
The logic for generating anticipated coefficients from delta varies with
glm family.
}
\keyword{internal}
