% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.simplex.R
\name{contr.simplex}
\alias{contr.simplex}
\title{Simplex Contrast Generator}
\usage{
contr.simplex(n, size = NULL)
}
\arguments{
\item{n}{The number of levels in the catagorical variable}

\item{size}{The length of the simplex vector. Default 1.}
}
\value{
A matrix of orthogonal contrasts.
}
\description{
Generates orthogonal contrasts. Each row is the vertex of an N-dimensional simplex. The only exception are contrasts for the 2-level case, which return 1 and -1.
}
\examples{
contr.simplex(4)
}
