% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectpower.R
\name{effectpower}
\alias{effectpower}
\title{Calculate Effect Power}
\usage{
effectpower(
  RunMatrix,
  levelvector,
  anticoef,
  alpha,
  vinv = NULL,
  degrees = NULL
)
}
\arguments{
\item{RunMatrix}{The model matrix}

\item{levelvector}{The number of levels in each parameter (1st is always the intercept)}

\item{anticoef}{The anticipated coefficients}

\item{alpha}{the specified type-I error}

\item{vinv}{The V inverse matrix}

\item{degrees}{Degrees of freedom}
}
\value{
The effect power for the parameters
}
\description{
Calculates the effect power given the anticipated coefficients and the type-I error
}
\keyword{internal}
