% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_anticoef.R
\name{gen_anticoef}
\alias{gen_anticoef}
\title{Generates Anticipated Coefficients}
\usage{
gen_anticoef(RunMatrix, model)
}
\arguments{
\item{RunMatrix}{The run matrix}
}
\value{
Anticipated coefficients.
}
\description{
Generates Anticipated Coefficients
}
\keyword{internal}
