% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_conservative_anticoef.R
\name{gen_conservative_anticoef}
\alias{gen_conservative_anticoef}
\title{Generates Default Conservative Anticipated Coefficients}
\usage{
gen_conservative_anticoef(modelmatrix)
}
\arguments{
\item{modelmatrix}{The model matrix}
}
\value{
Conservative Anticipated coefficients
}
\description{
Generates Default Conservative Anticipated Coefficients
}
\examples{
genparammatrix(6,2,1)
genparammatrix(6,2,1)
}

