% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_design.R
\name{gen_design}
\alias{gen_design}
\title{Generates optimal model matrix from factorial and model}
\usage{
gen_design(factorial, model, trials, optimality, contrastslist, repeats)
}
\arguments{
\item{factorial}{A factorial test matrix generated for all combinations of the factors in the model}

\item{model}{The model used to generate the test design}

\item{trials}{The number of trials}

\item{optimality}{The optimality criterion (e.g. "D")}

\item{contrastslist}{A list of the contrasts (e.g. "contr.sum")}

\item{repeats}{The number of times to repeat design generation the for best optimal condition}
}
\value{
The model matrix for the design, to be passed to eval_design
}
\description{
Creates design given a model and
desired number of runs, returning the model matrix. Currently
Used with eval_design/eval_design_mc to produce power estimations for designs.
Models with catagorical factors can only be linear and non-interacting.
}
\examples{
#generate all combinations of the factors with expand.grid()
factorial = expand.grid(a=c(-1,0,1),b=c(1,2,3,4),c=c(-1,1),d=c(-1,1))

#generate the design
designlinear = gen_design(factorial,~a+b+c+d,10,"D",100)

#generate the design with a quadratic model
designquadratic = gen_design(factorial,~(a+b+c+d)^2,15,"D",100)

mixedfactorial = expand.grid(a=as.factor(c(1,2,3,4,5,6)),b=c(1,0,-1),c=as.factor(c(1,2,3,4)),d=c(-1,1), stringsAsFactors = TRUE)
mixeddesign = gen_design(mixedfactorial,~a+b+c+d,15,"D",100)

#Evaluating the design can be done with eval_design and eval_design_mc (monte carlo)
}

