% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_noise_block.R
\name{generate_noise_block}
\alias{generate_noise_block}
\title{Generate Noise Block}
\usage{
generate_noise_block(noise, groups)
}
\arguments{
\item{RunMatrix}{The run matrix}
}
\value{
Normalized run matrix
}
\description{
Generates the noise to be added in the REML power calculation
}
\keyword{internal}
