% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genhypmatrix.R
\name{genhypmatrix}
\alias{genhypmatrix}
\title{Generate Hypothesis Matrix}
\usage{
genhypmatrix(parameters, levels, g)
}
\arguments{
\item{Parameters}{Number of parameters total in model}

\item{Levels}{Number of levels in parameter of interest}

\item{G}{Number of levels/parameters preceeding parameter of interest}
}
\value{
The parameter matrix L isolating the levels of parameter of interest
}
\description{
Generates hypothesis matrix for power calculation
}
\keyword{internal}
