% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genparammatrix.R
\name{genparammatrix}
\alias{genparammatrix}
\title{Generate Parameter Matrix}
\usage{
genparammatrix(parameters, levels, g)
}
\arguments{
\item{parameters}{Number of parameters total in model.}

\item{levels}{Number of levels in parameter of interest}

\item{g}{Number of levels/parameters preceeding parameter of interest}
}
\value{
The parameter vector Q isolating the levels of parameter of interest
}
\description{
Generates the parameter matrix L to isolate the levels of interest
in the calculation of power
}
\keyword{internal}
