% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterpower.R
\name{parameterpower}
\alias{parameterpower}
\title{Calculates parameter power}
\usage{
parameterpower(X, anticoef, alpha)
}
\arguments{
\item{X}{The model matrix}

\item{anticoef}{The anticipated coefficients}

\item{alpha}{the specified type-I error}
}
\value{
The parameter power for the parameters
}
\description{
Calculates parameter power
}
\examples{
mean(1)
}
\keyword{internal}

