% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterpower.R
\name{parameterpower}
\alias{parameterpower}
\title{Calculates parameter power}
\usage{
parameterpower(RunMatrix, anticoef, alpha, vInv = NULL)
}
\arguments{
\item{anticoef}{The anticipated coefficients}

\item{alpha}{the specified type-I error}

\item{vInv}{The V inverse matrix}

\item{X}{The model matrix}
}
\value{
The parameter power for the parameters
}
\description{
Calculates parameter power
}
\keyword{internal}
