% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterpower.R
\name{parameterpower}
\alias{parameterpower}
\title{Calculates parameter power}
\usage{
parameterpower(
  RunMatrix,
  levelvector = NULL,
  anticoef,
  alpha,
  vinv = NULL,
  degrees = NULL,
  parameter_names
)
}
\arguments{
\item{RunMatrix}{The run matrix}

\item{levelvector}{The number of levels in each parameter (1st is always the intercept)}

\item{anticoef}{The anticipated coefficients}

\item{alpha}{the specified type-I error}

\item{vinv}{The V inverse matrix}
}
\value{
The parameter power for the parameters
}
\description{
Calculates parameter power
}
\keyword{internal}
