% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducemodelmatrix.R
\name{reducemodelmatrix}
\alias{reducemodelmatrix}
\title{Remove columns not in model}
\usage{
reducemodelmatrix(ModelMatrix, model)
}
\arguments{
\item{ModelMatrix}{The model matrix}

\item{model}{The formula}
}
\value{
The reduced model matrix
}
\description{
Remove columns not in model
}
\examples{
mean(1)
}
\keyword{internal}

