% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_skpr_blockcols.R
\name{remove_skpr_blockcols}
\alias{remove_skpr_blockcols}
\title{Remove skpr-generated blocking columns}
\usage{
remove_skpr_blockcols(design)
}
\arguments{
\item{design}{The run matrix}
}
\value{
Data frame
}
\description{
Remove skpr-generated REML blocking columns if present
}
\keyword{internal}
