% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUI.R
\name{skprGUI}
\alias{skprGUI}
\title{Graphical User Interface for skpr}
\usage{
skprGUI(inputValue1, inputValue2)
}
\arguments{
\item{inputValue1}{It's the #1 input value.}

\item{inputValue2}{Runner up input value, participation award pending}
}
\description{
skprGUI provides a graphical user interface to skpr.
}
