% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUI.R
\name{skprGUI}
\alias{skprGUI}
\title{Graphical User Interface for skpr}
\usage{
skprGUI(browser = FALSE, return_app = FALSE)
}
\arguments{
\item{browser}{Default `FALSE`. Whether to open the application in an external browser.}

\item{return_app}{Default `FALSE`. If `TRUE`, this will return the shinyApp object.}

\item{inputValue1}{Required by Shiny}

\item{inputValue2}{Required by Shiny}
}
\description{
skprGUI provides a graphical user interface to skpr, within R Studio.
}
\examples{
#Type `skprGUI()` to begin

}
