% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skprGUIserver.R
\name{skprGUIserver}
\alias{skprGUIserver}
\title{Graphical User Interface for skpr}
\usage{
skprGUIserver(inputValue1, inputValue2)
}
\arguments{
\item{inputValue1}{Required by Shiny}

\item{inputValue2}{Required by Shiny}
}
\description{
skprGUI provides a graphical user interface to skpr, within R Studio.
}
\examples{
#Type skprGUIserver() to begin

\dontrun{skprGUIserver()}
}
